+++ avrCode/include/MILLIS.h
@@ -6,6 +6,6 @@ volatile unsigned long timer1_millis;
 ISR(TIMER1_COMPA_vect);
 void MillisInit(unsigned long f_cpu);
 unsigned long Millis(void);
-float GetDT();
-void SetDT(float _dt);
+double GetDT();
+void SetDT(double _dt);
 #endif
